Move the Export/Import layout buttons to the WidgetList
... might still rename it PanelList
